Separate out readme updates to dedicated test projects.
