+++ source/Quantities.TestTools/Quantities.TestTools.csproj
@@ -7,7 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Quantities\Quantities.csproj" PrivateAssets="All" />
+    <!-- ToDo: Re-enable PrivateAssets="All" -->
+    <ProjectReference Include="..\Quantities\Quantities.csproj" />
   </ItemGroup>
 
   <ItemGroup>
